# codeaware-py/openapi/current-release.json
{
  "components": {
    "schemas": {
...
      "AiReadmeCapability": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "reason": {
            "enum": [
              "available",
              "disabled",
              "roots_unavailable"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "reason"
        ],
        "title": "AiReadmeCapability",
        "type": "object"
      },
      "AiReadmeRequest": {
        "properties": {
          "project_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Project Name",
            "type": "string"
          },
          "project_path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Project Path",
            "type": "string"
          }
        },
        "required": [
          "project_name",
          "project_path"
        ],
        "title": "AiReadmeRequest",
        "type": "object"
      },
      "AiReadmeVO": {
        "properties": {
          "ai_model": {
            "default": "deepseek-v4-flash",
            "title": "Ai Model",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "snapshot_file_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot File Count"
          },
          "snapshot_generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Generated At"
          },
          "snapshot_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Hash"
          },
          "snapshot_truncated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Truncated"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "project_name",
          "title",
          "content"
        ],
        "title": "AiReadmeVO",
        "type": "object"
      },
...
      "ChatMessageVO": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessageVO",
        "type": "object"
      },
      "ChatRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "message": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponseVO": {
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "memory_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Summary"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ChatWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "conversation_id",
          "reply"
        ],
        "title": "ChatResponseVO",
        "type": "object"
      },
...
      "ConversationItem": {
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "conversation_id"
        ],
        "title": "ConversationItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Result_AiReadmeCapability_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiReadmeCapability"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AiReadmeCapability]",
        "type": "object"
      },
      "Result_AiReadmeVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiReadmeVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[AiReadmeVO]",
        "type": "object"
      },
      "Result_ChatResponseVO_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatResponseVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[ChatResponseVO]",
        "type": "object"
      },
...
      "Result_NoneType_": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[NoneType]",
        "type": "object"
      },
...
      "Result_Union_AiReadmeVO__NoneType__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiReadmeVO"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[Union[AiReadmeVO, NoneType]]",
        "type": "object"
      },
...
      "Result_list_ChatMessageVO__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessageVO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[list[ChatMessageVO]]",
        "type": "object"
      },
      "Result_list_ConversationItem__": {
        "properties": {
          "code": {
            "default": 1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConversationItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "msg": {
            "default": "success",
            "title": "Msg",
            "type": "string"
          }
        },
        "title": "Result[list[ConversationItem]]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CodeAware - AI 驱动的研发效能平台 (Python 重构)",
    "title": "codeaware",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ai-readme/capabilities": {
      "get": {
        "operationId": "capabilities_api_ai_readme_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AiReadmeCapability_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Capabilities",
        "tags": [
          "AIReadMe"
        ]
      }
    },
    "/api/ai-readme/generate": {
      "post": {
        "operationId": "generate_api_ai_readme_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiReadmeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_AiReadmeVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate",
        "tags": [
          "AIReadMe"
        ]
      }
    },
    "/api/ai-readme/{project_name}": {
      "get": {
        "operationId": "get_readme_api_ai_readme__project_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Union_AiReadmeVO__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Readme",
        "tags": [
          "AIReadMe"
        ]
      }
    },
...
    "/api/chat/conversations": {
      "get": {
        "operationId": "list_conversations_api_chat_conversations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_ConversationItem__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Conversations",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/conversations/{conversation_id}": {
      "delete": {
        "operationId": "delete_conversation_api_chat_conversations__conversation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Conversation",
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "description": "会话消息历史（PG 真相，按时间正序）。归属不匹配返回 404（不泄露存在性）。",
        "operationId": "get_conversation_api_chat_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_ChatMessageVO__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Conversation",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/send": {
      "post": {
        "description": "同步对话：drain TurnCoordinator -> ChatResponseVO。\n\nuser 经 DI 解析为 User（HTTP）；直连调用时为 Depends 实例或 None，跳过归属校验。",
        "operationId": "send_api_chat_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ChatResponseVO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_CONVERSATION_NOT_FOUND"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "conversation_id 不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_TURN_IN_PROGRESS"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "同一会话已有 turn 正在执行"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_REQUEST_INVALID"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "请求字段校验失败"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "CHAT_START_FAILED": {
                    "summary": "CHAT_START_FAILED",
                    "value": {
                      "code": 0,
                      "msg": "CHAT_START_FAILED"
                    }
                  },
                  "CONTEXT_FAILED": {
                    "summary": "CONTEXT_FAILED",
                    "value": {
                      "code": 0,
                      "msg": "CONTEXT_FAILED"
                    }
                  },
                  "MODEL_STREAM_FAILED": {
                    "summary": "MODEL_STREAM_FAILED",
                    "value": {
                      "code": 0,
                      "msg": "MODEL_STREAM_FAILED"
                    }
                  },
                  "PERSIST_FAILED": {
                    "summary": "PERSIST_FAILED",
                    "value": {
                      "code": 0,
                      "msg": "PERSIST_FAILED"
                    }
                  },
                  "POST_TURN_FAILED": {
                    "summary": "POST_TURN_FAILED",
                    "value": {
                      "code": 0,
                      "msg": "POST_TURN_FAILED"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "初始化失败或同步 Chat 核心失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/chat/send/stream": {
      "post": {
        "description": "流式对话：typed SSE。user 经 DI 解析；直连调用时跳过归属校验。",
        "operationId": "send_stream_api_chat_send_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "版本化 typed SSE Chat 事件流"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_CONVERSATION_NOT_FOUND"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "conversation_id 不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_TURN_IN_PROGRESS"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "同一会话已有 turn 正在执行"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_REQUEST_INVALID"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "请求字段校验失败"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "msg": "CHAT_START_FAILED"
                },
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            },
            "description": "SSE 响应建立前 Transaction A 初始化失败"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Stream",
        "tags": [
          "Chat"
        ]
      }
    },
...
  }
}